Restored constants folder and moved URL_SERVICE_RULES to enviroments
